Split out the Event type into its own module.
This conflicts with the existing Event type, so wait to replace it until that
type's usages are cleaned up.
